App/requirements.txt
fastapi==0.0.4
uvicorn==0.30.0
sqlalchemy==2.0.25
pydantic==2.7.2
jinja2==3.1.2
openai==1.51.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==23.0.0
alembic==1.13.2
//psycopg2==??